server/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
